Refactor: InputField ausgelagert
